app/requirements.txt
google-cloud-pubsub==1.0.0
google-cloud-storage==1.19.1
google-cloud-translate==1.6.0
google-cloud-vision==0.39.0
google-cloud-texttospeech==0.5.0
Pillow